WIP: move the Lookup state machine into a separate class
In preparation for some refactoring of the Lookup.run() logic
this change moves run(), and the various result getters
to a separate non-public class LookupHelper. This change
is not intended to introduce any behaviour changes.
